move header location
Seems to be working well.
However, platform-specific files that may include exports are not implemented. having them declared as export  may or may not be of interest.
