+++ config.h
@@ -1,12 +1,12 @@
-#define COREMAKE_STATIC
-#define COREMAKE_UNICODE
+//#define COREMAKE_STATIC
+//#define COREMAKE_UNICODE
 #define CONFIG_EBML_WRITING
-#define CONFIG_EBML_UNICODE
+//#define CONFIG_EBML_UNICODE
 #define HAVE_EBML2
 #define CONFIG_ZLIB
 #define CONFIG_MATROSKA2
-#define USE_PRECOMPILED_HEADERS
-#define CONFIG_DEBUGCHECKS
+//#define USE_PRECOMPILED_HEADERS
+//#define CONFIG_DEBUGCHECKS
 #define CONFIG_STDIO
 #define CONFIG_FILEPOS_64
 #define COREMAKE_CONFIG_HELPER
