+++ libebml2/ebmlnumber.c
@@ -28,7 +28,7 @@
 #include "ebml/ebml.h"
 #include "ebml/ebml_internal.h"
 
-#include "ebml2.h"
+#include "ebml/ebml2.h"
 
 static err_t ReadDataInt(ebml_integer *Element, stream *Input, const ebml_parser_context *ParserContext, bool_t AllowDummyElt, int Scope, size_t DepthCheckCRC)
 {
