+++ libebml2/ebmlstring.c
@@ -29,7 +29,7 @@
 #include "ebml/ebml.h"
 #include "ebml/ebml_internal.h"
 
-#include "ebml2.h"
+#include "ebml/ebml2.h"
 
 #if defined(CONFIG_EBML_UNICODE)
 err_t EBML_UniStringSetValue(ebml_string *Element,const tchar_t *Value)
